% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displacement_plot.R
\name{displacement_plot}
\alias{displacement_plot}
\title{Plot shoreline displacement curves}
\usage{
displacement_plot(
  target_curve = NA,
  displacement_curves = c("Horten", "Porsgrunn", "Tvedestrand", "Arendal"),
  greyscale = FALSE
)
}
\arguments{
\item{target_curve}{List holding one or more shoreline displacement curves.}

\item{displacement_curves}{Character vector specifying which geologically
informed displacement curves that should be plotted. Accepted values are
\code{c("Horten", "Porsgrunn", "Tvedestrand", "Arendal")}. All are included by
default.}

\item{greyscale}{Logical value indicating whether the plot should be in
greyscale or not. Defaults to FALSE.}
}
\value{
A plot displaying the underlying shoreline displacement curves and,
if provided, a target curve.
}
\description{
Function for plotting shoreline displacement curves. Calling to plot without
providing a target curve will display the four underlying geologically
derived displacement curves.
}
\examples{
# Empty plot for speed
displacement_plot(displacement_curves = "")

}
