% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten_list.R
\name{flatten_list}
\alias{flatten_list}
\title{Internal helper function for unpacking nested lists of plots}
\usage{
flatten_list(x)
}
\arguments{
\item{x}{Nested list of ggplot .}
}
\value{
List of ggplot objects
}
\description{
Utility function to extract a nested list of ggplots. Code taken from answers on Stackoverflow.
From user \@divibsan here:  https://stackoverflow.com/a/55733001 who in turn based their answer on user
\@Michael here: https://stackoverflow.com/a/41882883
}
\keyword{internal}
