% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_location_plot.R
\name{target_location_plot}
\alias{target_location_plot}
\title{Target location plot}
\usage{
target_location_plot(targets)
}
\arguments{
\item{targets}{Objects of class \code{sf} representing the sites to be dated. The first column beyond geom is taken as site name.}
}
\value{
A plot displaying the target locations relative to the shoreline isobases.
}
\description{
Function to plot the centroids of one or more sites to be dated, relative to the shoreline isobases of the employed displacement curves.
The basemap is a simplified representation of the coastline within the study area.
}
\examples{
target_point <- sf::st_sfc(sf::st_point(c(579570, 6582982)), crs = 32632)

target_location_plot(target_point)
}
