% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displacement_plot.R
\name{displacement_plot}
\alias{displacement_plot}
\title{Plot shoreline displacement curves}
\usage{
displacement_plot(interpolated_curve = NA)
}
\arguments{
\item{interpolated_curve}{List holding one or more interpolated shoreline displacement curves.}
}
\value{
A plot displaying the underlying shoreline displacement curves and, if provided, interpolated curves.
}
\description{
Function for plotting shoreline displacement curves. Calling to plot without providing interpolated curves will display the four underlying curves.
}
\examples{
# Create example point using the required coordinate system WGS84 UTM32N (EPSG: 32632).
target_pt <- sf::st_sfc(sf::st_point(c(579570, 6582982)), crs = 32632)

# Interpolate shoreline displacement curve to the target point location.
target_curve <- interpolate_curve(target_pt)

# Call to plot
displacement_plot(target_curve)
}
