% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_curve.R
\name{interpolate_curve}
\alias{interpolate_curve}
\title{Interpolate displacement curve using IDW}
\usage{
interpolate_curve(
  target,
  dispdat = load(system.file("extdata/displacement_curves.rda", package = "shoredate",
    mustWork = TRUE)),
  isobases = sf::st_read(system.file("extdata/isobases.gpkg", package = "shoredate"))
)
}
\arguments{
\item{target}{A spatial target location to where the new displacement curve is interpolated}

\item{dispdat}{Load existing displacement curves}

\item{isobases}{Load spatial lines representing the isobases of the existing displacement curves}
}
\value{
A data frame holding the interpolated displacement curve
}
\description{
Interpolate displacement curve using IDW
}
\examples{
target_pt <- sf::st_sfc(sf::st_point(c(579570, 6582982)), crs = 32632)
target_curve <- interpolate_curve(target_pt)
plot(target_curve$bce, target_curve$upperelev, type = "l")
lines(target_curve$bce, target_curve$lowerelev)
}
