% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shoredate_hdr.R
\name{shoredate_hdr}
\alias{shoredate_hdr}
\title{Find 95\% highest density region of shoreline date}
\usage{
shoredate_hdr(shorelinedate)
}
\arguments{
\item{shorelinedate}{A list of objects returned from shoreline_date().}
}
\value{
A data frame holding start and end points for segments of the 95\% highest density region
}
\description{
Function to find 95\% highest density region (HDR) for a provided shoreline date.
}
\examples{
# Create example point using the required coordinate system WGS84 UTM32N (EPSG: 32632).
target_pt <- sf::st_sfc(sf::st_point(c(579570, 6582982)), crs = 32632)

# Date target point, specifying the elevation instead of providing an elevation raster.
target_date <- shoreline_date(site = target_pt, elevation = 65)

# Retrieve and print HDR for the shoreline date
(shoredate_hdr(target_date))
}
