% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shoredate_plot.R
\name{shoredate_plot}
\alias{shoredate_plot}
\title{Plot shoreline date}
\usage{
shoredate_plot(
  shorelinedate,
  site_elevation = TRUE,
  displacement_curve = TRUE,
  lambda = TRUE,
  hdr_label = TRUE
)
}
\arguments{
\item{shorelinedate}{A list of objects returned from shoreline_date().}

\item{site_elevation}{Logical value indicating whether the site elevation and exponential decay function should be displayed. Default is TRUE.}

\item{displacement_curve}{Logical value indicating whether the displacement curve should be displayed. Default is TRUE.}

\item{lambda}{Logical value indicating whether the lambda value for the exponential decay function should be displayed. Default is TRUE.}

\item{hdr_label}{Logical value indicating whether the numeric values for the highest density regions should be displayed. Default is TRUE.}
}
\value{
A plot displaying the provided shoreline date.
}
\description{
Function for plotting shoreline date along with associated metadata.
}
\examples{
target_pt <- sf::st_sfc(sf::st_point(c(579570, 6582982)), crs = 32632)
target_date <- shoreline_date(site = target_pt, elevation = 65)
shoredate_plot(target_date)
}
\seealso{
\code{\link{shoreline_date}}
}
