% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shoredate_plot.R
\name{shoredate_plot}
\alias{shoredate_plot}
\title{Plot shoreline date}
\usage{
shoredate_plot(
  shorelinedate,
  site_elevation = TRUE,
  displacement_curve = TRUE,
  parameters = TRUE,
  isobase_direction = FALSE,
  highest_density_region = TRUE,
  hdr_label = TRUE,
  greyscale = FALSE
)
}
\arguments{
\item{shorelinedate}{A list of objects returned from shoreline_date().}

\item{site_elevation}{Logical value indicating whether the site elevation and gamma distribution should be displayed. Default is TRUE.}

\item{displacement_curve}{Logical value indicating whether the displacement curve should be displayed. Default is TRUE.}

\item{parameters}{Logical value indicating whether the parameters of the gamma distribution should be displayed. Default is TRUE.}

\item{isobase_direction}{Logical value indicating whether the direction of the isobases should be displayed. Default is FALSE.}

\item{highest_density_region}{Logical value indicating whether the 95\\% highest density region should be displayed. Defaults to TRUE.}

\item{hdr_label}{Logical value indicating whether the numeric values for the highest density regions rounded to the nearest 10 years should be displayed. Default is TRUE.}

\item{greyscale}{Logical value indicating whether the plot should be in greyscale or not. Defaults to FALSE.}
}
\value{
A plot displaying the provided shoreline date.
}
\description{
Function for plotting shoreline date along with associated metadata.
}
\examples{
target_point <- sf::st_sfc(sf::st_point(c(538310, 65442551)), crs = 32632)

target_date <- shoreline_date(site = target_point, elevation = 65,
                              isobase_direction = c(327,338))

shoredate_plot(target_date, isobase_direction = TRUE)
}
\seealso{
\code{\link{shoreline_date}}, \code{\link{annotate_custom}}
}
