% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shoreline_date.R
\name{shoreline_date}
\alias{shoreline_date}
\title{Shoreline date}
\usage{
shoreline_date(
  site,
  elev = NA,
  reso = 0.1,
  expratio = 0.168,
  elevavg = "mean",
  elevation = NA,
  interpolated_curve = NA
)
}
\arguments{
\item{site}{A spatial target representing the site to be dated}

\item{elev}{Elevation raster to be input if the elevation values are not provided manually}

\item{reso}{Numeric value specifying the resolution with which to step through the elevation distance between site and shoreline. Defaults to 0.1m}

\item{expratio}{Numeric value specifying the ratio with which the exponential function decays. Defaults to 0.168}

\item{elevavg}{Specified statistic to define elevation if this is to be derived from elevation raster}

\item{elevation}{Numeric elevation value to inform shoreline date unless}

\item{interpolated_curve}{Data frame holding shoreline displacement curve derived from interpolate_curve(). The interpolation function will be run if this is not provided-}
}
\description{
A function for shoreline dating a Stone Age site based on its present-day elevation and past shoreline displacement along the Norwegian Skagerrak coast.
}
\examples{
target_pt <- sf::st_sfc(sf::st_point(c(579570, 6582982)), crs = 32632)
target_date <- shoreline_date(site = target_pt, elevation = 65)
plot(target_date$years, target_date$probability, type = "l")
}
