% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shoreline_date.R
\name{shoreline_date}
\alias{shoreline_date}
\title{Shoreline date}
\usage{
shoreline_date(
  site,
  elev_raster = NA,
  elev_reso = 0.01,
  cal_reso = 10,
  isobase_direction = 327,
  model_parameters = c(0.286, 0.048),
  elevavg = "mean",
  elevation = NA,
  interpolated_curve = NA,
  normalise = TRUE,
  sparse = FALSE
)
}
\arguments{
\item{site}{Vector giving site name, or, if displacement curve is to be interpolated, an object of class \code{sf} representing the site to be dated.}

\item{elev_raster}{Elevation raster to be input if the elevation values are not provided manually.}

\item{elev_reso}{Numeric value specifying the resolution with which to step through the elevation distance between site and shoreline. Defaults to 0.01m.}

\item{cal_reso}{Numeric value specifying the resolution to use on the calendar scale. Defaults to 10.}

\item{isobase_direction}{A single numeric value or a vector of values defining the direction(s) of the isobases. Defaults to 327.}

\item{model_parameters}{Vector of numeric values specifying the shape \eqn{\alpha} and rate \eqn{\sigma} of the gamma distribution. Defaults to \eqn{\alpha} = 0.286 and \eqn{\sigma} = 0.048.}

\item{elevavg}{Specified statistic to define elevation if this is to be derived from elevation raster. Defaults to mean.}

\item{elevation}{Numeric elevation value to inform shoreline date unless an elevation raster is provided.}

\item{interpolated_curve}{List holding shoreline displacement curve. interpolate_curve() will be run if this is not provided.}

\item{normalise}{Logical value specifying whether the shoreline date should be normalised to sum to unity. Defaults to TRUE.}

\item{sparse}{Logical value specifying if only site name and shoreline date should be returned. Defaults to FALSE.}
}
\value{
A list containing the shoreline date and associated parameters.
}
\description{
A function for shoreline dating a Stone Age site based on its present-day elevation, it's likely elevation above sea-level when in use and the trajectory of past shoreline displacement on the Norwegian Skagerrak coast.
}
\examples{
# Create example point using the required coordinate system WGS84 UTM32N (EPSG: 32632).
target_point <- sf::st_sfc(sf::st_point(c(538310, 65442551)), crs = 32632)

# Date target point, manually specifying the elevation instead of providing an elevation raster.
target_date <- shoreline_date(site = target_point, elevation = 46)

# Call to plot
shoredate_plot(target_date)
}
