% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shoreline_date.R
\name{shoreline_date}
\alias{shoreline_date}
\title{Shoreline date}
\usage{
shoreline_date(
  sites,
  elevation = NA,
  elev_raster = NA,
  elev_reso = 0.01,
  cal_reso = 10,
  isobase_direction = 327,
  model_parameters = c(0.286, 0.048),
  elevavg = "mean",
  interpolated_curve = NA,
  hdr_prob = 0.95,
  normalise = TRUE,
  sparse = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{sites}{Vector giving one or more site names, or, if displacement curves
are to be interpolated, objects of class \code{sf} representing the sites to be
dated. In the case of a spatial geometry, the first column is taken as
site name.}

\item{elevation}{Vector of numeric elevation values to inform shoreline dates
unless an elevation raster is provided.}

\item{elev_raster}{Elevation raster to be input if the elevation values are
not provided manually.}

\item{elev_reso}{Numeric value specifying the resolution with which to step
through the elevation distance between site and shoreline.
Defaults to 0.001m.}

\item{cal_reso}{Numeric value specifying the resolution to use on the
calendar scale. Must be a power of 10. Defaults to 10.}

\item{isobase_direction}{A vector of numeric values defining the direction(s)
of the isobases. Defaults to 327.}

\item{model_parameters}{Vector of two numeric values specifying the shape and
scale of the gamma distribution. Defaults to c(0.286, 0.048), denoting the
shape and scale, respectively.}

\item{elevavg}{Specified statistic to define elevation if this is to be
derived from an elevation raster. Defaults to mean.}

\item{interpolated_curve}{List holding shoreline displacement curve.
\code{interpolate_curve()} will be run if this is not provided.}

\item{hdr_prob}{Numeric value specifying the coverage of the highest density
region. Defaults to 0.95.}

\item{normalise}{Logical value specifying whether the shoreline date should
be normalised to sum to unity. Defaults to TRUE.}

\item{sparse}{Logical value specifying if only site name and shoreline date
should be returned. Defaults to FALSE.}

\item{verbose}{Logical value indicating whether progress should be printed to
console. Defaults to FALSE.}
}
\value{
A nested list of class \code{shoreline_date} holding the shoreline date
results and associated metadata for each dated site for each isobase
direction. The elements of each date is:
\itemize{
\item The name of the site, \code{site_name}.
\item The elevation of the site, \code{site_elev}.
\item A data frame, \code{date} , with the columns \code{bce} where negative values
indicate years BCE and positive CE, and \code{probability}, which gives the
probability mass for each year.
\item The start values for the HDR ranges, \code{hdr_start}.
\item The end values for the HDR ranges, \code{hdr_end}.
\item The probability level for the HDR, \code{hdr_prob}.
\item A data frame, \code{dispcurve}, holding the displacement curve used for dating
the site. This has the columns \code{bce}, giving years BCE/CE. \code{lowerelev},
the lower limit for the elevation of the shoreline for each year.
\code{upperelev}, the upper limit for elevation of the shoreline for each year.
\item The direction of the isobases in use if the displacement curve has been
interpolated, \code{dispcurve_direction}.
\item The parameters for the gamma distribution, \code{dispcurve_direction}. The
first value gives the shape and the second value the scale of the
distribution.
\item A data frame holding the gamma distribution, \code{gammdat}. The column
\code{offset} denotes the vertical distance (m) from the shoreline, as specified
by the \code{elev_reso} argument. \code{px} is the cumulative probability at each step
of \code{offset}, and \code{probs} is the probability of each step found by
subtracting the preceeding value from each value of \code{px}.
\item The resolution on the calendar scale, \code{cal_reso}.
}
}
\description{
A function for shoreline dating Stone Age sites based on their present-day
elevation, their likely elevation above sea-level when in use and the
trajectory of past shoreline displacement on the Norwegian Skagerrak coast.
}
\examples{
# Create two example points
target_points <- sf::st_sfc(sf::st_point(c(538310, 6544255)),
  sf::st_point(c(572985, 6563115)))

# Set these to the required coordinate system WGS84 UTM32N (EPSG: 32632)
target_points <- sf::st_set_crs(target_points, 32632)

# Date target points, manually specifying the elevations instead of providing
# an elevation raster and print the results to console.
shoreline_date(sites = target_points, elevation = c(46, 60))
}
