% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shoreline_date.R
\name{shoreline_date}
\alias{shoreline_date}
\title{Shoreline date}
\usage{
shoreline_date(
  sites,
  elevation = NA,
  elev_raster = NA,
  elev_reso = 0.01,
  cal_reso = 10,
  isobase_direction = 327,
  model_parameters = c(0.286, 0.048),
  elevavg = "mean",
  interpolated_curve = NA,
  normalise = TRUE,
  sparse = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{sites}{Vector giving one or more site names, or, if displacement curves are to be interpolated, objects of class \code{sf} representing the sites to be dated. In the case of a spatial geometry, the first column is taken as site name.}

\item{elevation}{Vector of numeric elevation values to inform shoreline dates unless an elevation raster is provided.}

\item{elev_raster}{Elevation raster to be input if the elevation values are not provided manually.}

\item{elev_reso}{Numeric value specifying the resolution with which to step through the elevation distance between site and shoreline. Defaults to 0.001m.}

\item{cal_reso}{Numeric value specifying the resolution to use on the calendar scale. Must be a power of 10. Defaults to 10.}

\item{isobase_direction}{A vector of numeric values defining the direction(s) of the isobases. Defaults to 327.}

\item{model_parameters}{Vector of numeric values specifying the shape and scale of the gamma distribution. Defaults to shape = 0.286 and scale = 0.048.}

\item{elevavg}{Specified statistic to define elevation if this is to be derived from elevation raster. Defaults to mean.}

\item{interpolated_curve}{List holding shoreline displacement curve. interpolate_curve() will be run if this is not provided.}

\item{normalise}{Logical value specifying whether the shoreline date should be normalised to sum to unity. Defaults to TRUE.}

\item{sparse}{Logical value specifying if only site name and shoreline date should be returned. Defaults to FALSE.}

\item{verbose}{Logical value indicating whether progress should be printed to console. Defaults to FALSE.}
}
\value{
A list of class shoredates holding the shoreline date results and associated parameters.
}
\description{
A function for shoreline dating Stone Age sites based on their present-day elevation, their likely elevation above sea-level when in use and the trajectory of past shoreline displacement on the Norwegian Skagerrak coast.
}
\examples{
# Create example points.
target_points <- sf::st_sfc(sf::st_point(c(538310, 6544255)),
 sf::st_point(c(572985, 6563115)))

# Set these to the required coordinate system WGS84 UTM32N (EPSG: 32632).
target_points <- sf::st_set_crs(target_points, 32632)

# Date target points, manually specifying the elevations instead of providing
# an elevation raster.
target_dates <- shoreline_date(sites = target_points, elevation = c(46, 60))

# Call to plot.
shoredate_plot(target_dates)
}
